feat: decoder
