+++ siopv2/Cargo.toml
@@ -25,7 +25,6 @@ url = { version = "2.3.1", features = ["serde"] }
 is_empty = "0.2.0"
 serde_urlencoded = "0.7.1"
 derive_more = "0.99.16"
-bimap = "0.6"
 
 [dev-dependencies]
 ed25519-dalek = "1.0.1"
