+++ siopv2/src/authentication/subject.rs
@@ -6,7 +6,6 @@ use std::{str::FromStr, sync::Arc};
 /// This [`Subject`] trait is used to sign and verify JWTs.
 pub trait Subject: Sign + Verify {
     fn identifier(&self) -> Result<String>;
-    // TODO: Remove?
     fn type_(&self) -> Result<SubjectSyntaxType> {
         SubjectSyntaxType::from_str(&self.identifier()?)
     }
