+++ siopv2/src/lib.rs
@@ -2,6 +2,7 @@ pub mod authentication;
 pub mod claims;
 pub mod client_metadata;
 pub mod collection;
+pub mod decoder;
 pub mod jwt;
 pub mod provider;
 pub mod relying_party;
@@ -20,6 +21,7 @@ pub use authentication::{
 pub use claims::{ClaimRequests, StandardClaimsRequests, StandardClaimsValues};
 pub use client_metadata::ClientMetadata;
 pub use collection::Collection;
+pub use decoder::Decoder;
 pub use jwt::JsonWebToken;
 pub use provider::Provider;
 pub use relying_party::RelyingParty;
